Update README.md
Questions preceded by a * are optional. Although they can be skipped without any deduction, it is important to know and understand the results they contain.

Ex. 1 — Perfect matching in a bipartite graph

Let G = V , E be a bipartite graph where V = L ∪ R. A perfect matching in G is a subset of E where every vertex is contained in exactly one edge. Let A be the matrix whose rows correspond to vertices in L and columns to vertices in R. Each element ai ,j of A is defined as a variable Xi ,j , if vertices i and j are connected and 0 otherwise.

    Expressing the determinant of A as a polynomial prove that it is identically zero if and only if G has no perfect matching.

    Deduce an algorithm to decide if a bipartite graph has a perfect matching.

    What are the complexity and error probability of this algorithm?

    As deterministic polynomial time algorithms already exist, discuss the usefulness of this strategy.